data_forwarding/requirements.txt
pyserial ~= 3.5
paho-mqtt ~= 2.1.0